Update kotlin-code-style.md
I think Sindre was right from the beginning, the externalisation stuff was too dependent on context. I also place companion stuff before, since we say it should be first in the call. 
After some experiments on the extension stuff, I think it's not useful to have `Extension` in the file name.
For the Unit part, what do you think